% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{fit}
\alias{fit}
\title{Fit Model to Log-Returns via MLE}
\usage{
fit(object, log_returns, ...)
}
\arguments{
\item{object}{A \linkS4class{JumpDiffModel} object.}

\item{log_returns}{Numeric vector of observed log returns.}

\item{...}{Passed to \code{\link[stats]{optim}}.}
}
\value{
A \linkS4class{JDFitResult} object.
}
\description{
Fit Model to Log-Returns via MLE
}
