% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merton_model.R, R/merton_fit.R
\docType{class}
\name{JDFitResult-class}
\alias{JDFitResult-class}
\alias{show,JDFitResult-method}
\alias{confint,JDFitResult-method}
\title{JDFitResult: MLE Estimation Output Class}
\usage{
\S4method{show}{JDFitResult}(object)

\S4method{confint}{JDFitResult}(object, parm, level = 0.95, ...)
}
\arguments{
\item{object}{A \linkS4class{JDFitResult} object.}

\item{parm}{Ignored (all parameters returned).}

\item{level}{Confidence level (default 0.95).}

\item{...}{Unused.}
}
\description{
JDFitResult: MLE Estimation Output Class
}
\section{Methods (by generic)}{
\itemize{
\item \code{confint(JDFitResult)}: Wald-type 95\\% confidence intervals.

}}
\section{Slots}{

\describe{
\item{\code{estimates}}{Named numeric. Parameter estimates.}

\item{\code{se}}{Named numeric. Standard errors.}

\item{\code{loglik}}{Numeric. Log-likelihood at optimum.}

\item{\code{converged}}{Logical. Did optimisation converge?}

\item{\code{model_type}}{Character. Model name.}
}}

