% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merton_model.R
\docType{class}
\name{JDSimResult-class}
\alias{JDSimResult-class}
\title{JDSimResult: Simulation Output Class}
\description{
JDSimResult: Simulation Output Class
}
\section{Slots}{

\describe{
\item{\code{paths}}{Matrix. Simulated price paths (n x steps+1).}

\item{\code{times}}{Numeric vector. Time grid.}

\item{\code{params}}{List. Model parameters used in simulation.}

\item{\code{model_type}}{Character. Model name.}
}}

