% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{jumpMoments}
\alias{jumpMoments}
\title{Theoretical Moments of the Log-Return Distribution}
\usage{
jumpMoments(object)
}
\arguments{
\item{object}{A \linkS4class{JumpDiffModel} object.}
}
\value{
Named numeric vector: mean, variance, skewness, kurtosis.
}
\description{
Theoretical Moments of the Log-Return Distribution
}
