% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{loglik}
\alias{loglik}
\title{Compute Negative Log-Likelihood}
\usage{
loglik(object, log_returns, ...)
}
\arguments{
\item{object}{A \linkS4class{JumpDiffModel} object.}

\item{log_returns}{Numeric vector of log asset returns.}

\item{...}{Passed to methods.}
}
\value{
Scalar negative log-likelihood value.
}
\description{
Compute Negative Log-Likelihood
}
