% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merton_model.R
\docType{class}
\name{MertonModel-class}
\alias{MertonModel-class}
\alias{show,MertonModel-method}
\title{MertonModel: Merton (1976) Jump-Diffusion Model}
\usage{
\S4method{show}{MertonModel}(object)
}
\arguments{
\item{object}{A \linkS4class{MertonModel} object.}
}
\description{
Extends \linkS4class{JumpDiffModel} with log-normal jump parameters.
}
\section{Slots}{

\describe{
\item{\code{mu_j}}{Numeric. Mean log-jump size.}

\item{\code{sigma_j}}{Positive numeric. Std dev of log-jump sizes.}
}}

\references{
Merton, R.C. (1976). Option pricing when underlying stock returns are
discontinuous. \emph{Journal of Financial Economics}, 3(1-2), 125-144.
}
